requirements.txt
h5py==2.10.0
image==1.5.32
imageio==2.8.0
matplotlib==3.1.3
numpy==1.19.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
Pillow==7.1.2
PyExifTool==0.1.1
pytorch-lightning==0.8.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
spectral==0.21
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
wandb==0.10.5
seaborn==0.11.0
